ChickenChecker-1314-main/ChickenChecker-1314-main/requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==4.0.2
django-adaptors==0.2.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fonttools==4.29.1
imageio==2.16.1
ipykernel==6.7.0
ipython==8.2.0
jedi==0.18.1
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mysqlclient==2.1.0
nest-asyncio==1.5.5
numpy==1.22.2
packaging==21.3
pandas==1.4.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
reportlab==3.6.9
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5
